Rename JS files without component exports
